+++ src/main/java/it/polpetta/libris/contract/AbstractURLSearcher.java
@@ -1,0 +1,8 @@
+package it.polpetta.libris.contract;
+
+/**
+ * Created by zanna on 08/05/17.
+ * its a template method
+ */
+public class AbstractURLSearcher {
+}
