+++ src/main/java/it/polpetta/libris/image/azure/contract/IAzureImageSearchResult.java
@@ -1,13 +1,13 @@
-package it.polpetta.libris.utils.SearchResult.googleImageSearch;
+package it.polpetta.libris.image.azure.contract;
 
-import it.polpetta.libris.utils.SearchResult.ISearchResult;
+import it.polpetta.libris.contract.ISearchResult;
 
 import java.util.ArrayList;
 
 /**
  * Created by davide on 06/05/17.
  */
-public interface IGoogleImageSearchResult extends ISearchResult {
+public interface IAzureImageSearchResult extends ISearchResult {
 
     String getBestGuess();
     ArrayList<String> getLinks();
