+++ src/main/java/it/polpetta/libris/image/azure/imageRecognition/AzureImageRecognitionQueryBuilder.java
@@ -1,11 +1,11 @@
-package it.polpetta.libris.azure.imageRecognition;
+package it.polpetta.libris.image.azure.imageRecognition;
 
-import it.polpetta.libris.IQuery;
-import it.polpetta.libris.IQueryBuilder;
-import it.polpetta.libris.azure.imageRecognition.searchers.URLFactoryMethodSearcher;
-import it.polpetta.libris.google.imageSearch.Coordinates;
-import it.polpetta.libris.utils.ISearcher;
-import it.polpetta.libris.utils.SearchResult.ISearchResult;
+import it.polpetta.libris.contract.IQuery;
+import it.polpetta.libris.contract.IQueryBuilder;
+import it.polpetta.libris.image.azure.imageRecognition.searchers.URLFactoryMethodSearcher;
+import it.polpetta.libris.utils.Coordinates;
+import it.polpetta.libris.contract.ISearcher;
+import it.polpetta.libris.contract.ISearchResult;
 
 import java.io.File;
 import java.net.URL;
