+++ src/main/java/it/polpetta/libris/image/contract/AbstractURLImageSearcher.java
@@ -1,0 +1,10 @@
+package it.polpetta.libris.image.contract;
+
+import it.polpetta.libris.contract.AbstractURLSearcher;
+
+/**
+ * Created by zanna on 08/05/17.
+ * its a template method
+ */
+public class AbstractURLImageSearcher extends AbstractURLSearcher {
+}
