+++ src/main/java/it/polpetta/libris/image/contract/IImageQueryBuilder.java
@@ -1,0 +1,9 @@
+package it.polpetta.libris.image.contract;
+
+import it.polpetta.libris.contract.IQueryBuilder;
+
+/**
+ * Created by zanna on 08/05/17.
+ */
+public interface IImageQueryBuilder extends IQueryBuilder {
+}
