+++ src/main/java/it/polpetta/libris/image/contract/IImageSearchResult.java
@@ -1,0 +1,7 @@
+package it.polpetta.libris.image.contract;
+
+/**
+ * Created by zanna on 08/05/17.
+ */
+public interface IImageSearchResult {
+}
