+++ src/main/java/it/polpetta/libris/image/google/searchers/URLFactoryMethodSearcher.java
@@ -1,9 +1,8 @@
-package it.polpetta.libris.google.imageSearch.searchers;
+package it.polpetta.libris.image.google.searchers;
 
-import it.polpetta.libris.google.imageSearch.Coordinates;
-import it.polpetta.libris.google.imageSearch.IGoogleImageSearcher;
-import it.polpetta.libris.utils.AbstractFactoryMethodSearcher;
-import it.polpetta.libris.utils.ISearcher;
+import it.polpetta.libris.utils.Coordinates;
+import it.polpetta.libris.image.google.IGoogleImageSearcher;
+import it.polpetta.libris.contract.AbstractFactoryMethodSearcher;
 
 import java.net.URL;
