+++ src/main/java/it/polpetta/libris/image/google/searchers/URLSearcher.java
@@ -1,11 +1,11 @@
-package it.polpetta.libris.google.imageSearch.searchers;
+package it.polpetta.libris.image.google.searchers;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
-import it.polpetta.libris.google.imageSearch.Coordinates;
-import it.polpetta.libris.google.imageSearch.IGoogleImageSearcher;
+import it.polpetta.libris.utils.Coordinates;
+import it.polpetta.libris.image.google.IGoogleImageSearcher;
 import it.polpetta.libris.utils.SearchResult.SearchResultBuilderAbstractFactory;
-import it.polpetta.libris.utils.SearchResult.googleImageSearch.IGoogleImageSearchResult;
+import it.polpetta.libris.image.google.contract.IGoogleImageSearchResult;
 import org.jsoup.nodes.Document;
 import org.jsoup.nodes.Element;
 import org.jsoup.parser.Parser;
